feat: user 패키지 생성
