+++ src/main/java/com/mars/entity/Order.java
@@ -1,5 +1,6 @@
 package com.mars.entity;
 
+import com.mars.user.domain.entity.User;
 import java.util.ArrayList;
 import java.util.List;
 import javax.persistence.CascadeType;
