+++ src/main/java/com/mars/user/applicaton/service/UserService.java
@@ -1,0 +1,5 @@
+package com.mars.user.applicaton.service;
+
+public class UserService {
+
+}
