+++ src/main/java/com/mars/user/applicaton/vo/UserVo.java
@@ -1,0 +1,5 @@
+package com.mars.user.applicaton.vo;
+
+public class UserVo {
+
+}
