+++ src/main/java/com/mars/user/domain/repo/UserRepository.java
@@ -1,0 +1,5 @@
+package com.mars.user.domain.repo;
+
+public interface UserRepository {
+
+}
