+++ src/main/java/com/mars/user/domain/vo/UserVo.java
@@ -1,0 +1,5 @@
+package com.mars.user.domain.vo;
+
+public class UserVo {
+
+}
