+++ src/main/java/com/mars/user/infra/UserRepositoryImpl.java
@@ -1,0 +1,5 @@
+package com.mars.user.infra;
+
+public class UserRepositoryImpl {
+
+}
