+++ src/main/java/com/mars/user/presentation/controller/UserController.java
@@ -1,0 +1,5 @@
+package com.mars.user.presentation.controller;
+
+public class UserController {
+
+}
