+++ src/main/java/com/mars/user/presentation/dto/UserDto.java
@@ -1,0 +1,5 @@
+package com.mars.user.presentation.dto;
+
+public class UserDto {
+
+}
